% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_slopeDif.R
\name{plot.slopetest}
\alias{plot.slopetest}
\alias{plot}
\title{Create an interaction plot from an object of class "slopetest".}
\usage{
plot(x, legLoc = "topright", ...)
}
\arguments{
\item{x}{An object of class "slopetest". The result of an \code{slopeDif()} function.}

\item{legLoc}{A string. Where should the legend be located? Default = "topright".}

\item{...}{Additional arguments passed to \code{plot()}.}
}
\value{
A three-way interaction plot
}
\description{
Creates an interaction plot comparing high and low values of X at high and low value combinations of Z and W.
}
\examples{
library(mixedPowerSim)
library(lme4)

nCluster <- 20
nObs <- 15
db0 <- .1
db1 <- .1
db2 <- .1
db3 <- .1
db4 <- .1
db5 <- .1
db6 <- .1
db7 <- .1
SDb0 <- 1
SDresid <- 5

simDat <- simMixedXYZ(nCluster = nCluster, nObs = nObs,
 db0 = db0, db1 = db1, db2 = db2,
 db3 = db3, db4 = db4, db5 = db5, db6 = db6, db7 = db7,
 SDb0 = SDb0, SDresid = SDresid, XWithin = TRUE, ZWithin = FALSE, WWithin = TRUE)

 mod <- lmer(Y~X * Z * W + (1|Cluster), data = simDat$simulatedData)

 difTest <- slopeDif(mod)

 plot(difTest)
 ## OR
 plot(difTest)
}
