% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopeDif.R
\name{slopeDif}
\alias{slopeDif}
\title{Conduct a slope difference test for three-way interaction mixed-effects models.}
\usage{
slopeDif(x)
}
\arguments{
\item{x}{An object of class "lmerMod". The result of an \code{lmer()} function with a random intercept.}
}
\value{
An object of class "slopetest" with the following components:
\itemize{
\item{Expected_Y: }{A data frame with the expected values of Y at high and low values of X, Z, and W.}
\item{SlopeTests: }{A data frame representing slope tests as defined in Dawson and Richter (2006).}
}
}
\description{
Calculates the expected value of the outcome variable (Y) at all combinations of high and low values (+/- 1SD) of X, Z, and W.
Additionally, this function calculates the slope differences between all slope pairs.
}
\examples{
library(mixedPowerSim)
library(lme4)

nCluster <- 20
nObs <- 15
db0 <- .1
db1 <- .1
db2 <- .1
db3 <- .1
db4 <- .1
db5 <- .1
db6 <- .1
db7 <- .1
SDb0 <- 1
SDresid <- 5

simDat <- simMixedXYZ(nCluster = nCluster, nObs = nObs,
 db0 = db0, db1 = db1, db2 = db2,
 db3 = db3, db4 = db4, db5 = db5, db6 = db6, db7 = db7,
 SDb0 = SDb0, SDresid = SDresid, XWithin = TRUE, ZWithin = FALSE, WWithin = TRUE)

 mod <- lmer(Y~X * Z * W + (1|Cluster), data = simDat)

 slopeDif(mod)
}
