% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XYZpowerSim.R
\name{XYZpowerSim}
\alias{XYZpowerSim}
\title{Run a power analysis of slope differences.}
\usage{
XYZpowerSim(R = 1000, clusters = NA, obs =  NA, sigLevel = .05 ...)
}
\arguments{
\item{R}{Integer. Number of replications within each cell of simulation.}

\item{clusters}{A vector of integers. Number of clusters to be crossed with \code{obs} e.g., \code{c(5,10,15)}.}

\item{obs}{A vector of integers. Number of observations for each level of to be crossed with \code{clusters} e.g., \code{c(5,10,15)}.}

\item{sigLevel}{Numeric. $alpha$ value for determine significant p-values in each simulation.}

\item{...}{Arguments passed to \link{simMixedXYZ}.}
}
\value{
A data frame of simulated parameters.
}
\description{
Creates a simulated data set based upon the supplied arguments.
}
\details{
By specifying effect sizes (\code{db0}, ..., \code{db7}) of betas as seen in Dawson and Richter (2006),
number of clusters, and number of observations within cluster, this function creates a data set with the
appropriate relationships. Variables X, Y, and Z are each drawn from N~(0,1).
}
\examples{
library(mixedPowerSim)
nCluster <- c(5, 10, 15)
nObs <- c(5, 10, 15)
reliability <- .8
db0 <- .1
SDb0 <- 1
SDresid <- 5
XYZpowerSim(R = 3, clusters = nCluster, obs =  nObs, sigLevel = .05, type = "slopeDifference",
b0 = db0, a_Dif = 0, b_Dif = 0,c_Dif = 0,
d_Dif = 0, e_Dif = 0, f_Dif = 0,
SDb0 = SDb0, SDresid = SDresid, XWithin = TRUE, ZWithin = FALSE, WWithin = TRUE, reliability = reliability)
}
