% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMixedXYZ.R
\name{simMixedXYZ}
\alias{simMixedXYZ}
\title{Simulate data for a three-way interaction of nested observations.}
\usage{
simMixedXYZ(nCluster, nObs,
 db0, db1, db2, db3, db4, db5, db6, db7,
 SDb0, SDresid, XWithin = TRUE, ZWithin = TRUE, WWithin = TRUE)
}
\arguments{
\item{nCluster}{Integer. Number of clusters (level-2 observations).}

\item{nObs}{Integer. Number of observations within each cluster.}

\item{db0}{Numeric. Effect size of b0.}

\item{db1}{Numeric. Effect size of b1.}

\item{db2}{Numeric. Effect size of b2.}

\item{db3}{Numeric. Effect size of b3.}

\item{db4}{Numeric. Effect size of b4.}

\item{db5}{Numeric. Effect size of b5.}

\item{db6}{Numeric. Effect size of b6.}

\item{db7}{Numeric. Effect size of b7.}

\item{SDb0}{Numeric. Between cluster standard deviation of b0.}

\item{SDresid}{Numeric. Standard deviation of residuals.}

\item{XWithin}{Logical. If TRUE, variable X is simulated as a within-cluster variable. If FALSE, variable X is simulated as a between-cluster variable.}

\item{ZWithin}{Logical. If TRUE, variable Z is simulated as a within-cluster variable. If FALSE, variable Z is simulated as a between-cluster variable.}

\item{WWithin}{Logical. If TRUE, variable W is simulated as a within-cluster variable. If FALSE, variable W is simulated as a between-cluster variable.}
}
\value{
A data frame with the following columns:
\itemize{
\item{Cluster:}{ Cluster identification number.}
\item{Observation:}{ Observation number within cluster.}
\item{Y:}{ Outcome variable.}
\item{X:}{ X variable.}
\item{Z:}{ Z variable.}
\item{W:}{ W variable.}
}
}
\description{
Creates a simulated data set based upon the supplied arguments.
}
\details{
By specifying effect sizes (\code{db0}, ..., \code{db7}) of betas as seen in Dawson and Richter (2006),
number of clusters, and number of observations within cluster, this function creates a data set with the
appropriate relationships. Variables X, Y, and Z are each drawn from N~(0,1).
}
\examples{
library(mixedPowerSim)
nCluster <- 20
nObs <- 15
db0 <- .1
db1 <- .1
db2 <- .1
db3 <- .1
db4 <- .1
db5 <- .1
db6 <- .1
db7 <- .1
SDb0 <- 1
SDresid <- 5
simDat <- simMixedXYZ(nCluster = nCluster, nObs = nObs,
 db0 = db0, db1 = db1, db2 = db2,
 db3 = db3, db4 = db4, db5 = db5, db6 = db6, db7 = db7,
 SDb0 = SDb0, SDresid = SDresid, XWithin = TRUE, ZWithin = FALSE, WWithin = TRUE)
}
